+++ crates/proto/src/version/v800/types/biome_mesa_surface.rs
@@ -3,9 +3,9 @@ use bedrockrs_macros::ProtoCodec;
 #[derive(ProtoCodec, Debug, Clone, Copy)]
 pub struct BiomeMesaSurfaceData {
     #[endianness(le)]
-    pub clay_material: i64,
+    pub clay_material: i32,
     #[endianness(le)]
-    pub hard_clay_material: i64,
+    pub hard_clay_material: i32,
     pub bryce_pillars: bool,
     pub has_forest: bool,
 }
